Update README regarding subteams and leads
Per discussion between Skander and me, I will be the lead of the
NotificationSystem subteam this year.  (We are both already on the
CODEOWNERS list, along with Scott and Michelle, but I'm leading
the actual work.)

Also tentatively removed mention of the ModeratorFrontEnd and
UserExperience subteams per https://github.com/orcasound/orcasite/issues/917

If there is another qualified volunteer to lead the Infrastructure subteam
I will happily hand over the lead on that one.

Not sure if there are other subteam/lead updates to make before the 2025 hackathon.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>