Let dependabot track nuget dependencies
There are a bunch of code scanning issues found in this repository
under the Security tab.  Most appear to be due to outdated nuget
packages.  Dependabot already tracks github actions, but not
nuget package dependencies.

As the new lead for the NotificationSystem subsystem (and the old
lead for github infrastructure), I want to enable them on at least
the NotificationSystem subsystem.  If others don't want them on other
subsystems, I can update this PR to restrict the directory to only
NotificationSystem.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>