Add Andrews Bay hydrophone
In preparation for new hydrophone deployment at Andrews Bay within the
next month or so.

Currently the RPI is streaming silence to the
dev-streaming-orcasound.net bucket instead of the audio-orcasound-net
bucket.  The bucket will need to change to audio-orcasound-net before
this PR can be successfully tested.  Will leave in Draft until then.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>