Create a github workflow for testing InferenceSystem (#302)
* Let dependabot check InferenceSystem dependencies

Some current dependencies have problems in that InferenceSystem
scripts relies on some functionality that has been removed in
later versions, so requirements.txt must currently pin such
dependencies to an older version, which this PR does for
numpy and spacy.  Without that fix, the README instructions
fail.

As noted in issue #211 there is currently no notification when
there is a problem like this since there's no CI workflows for
the InferenceSystem yet, and no dependabot checks.  This PR
adds dependabot checks.  A separate PR will add a CI workflow.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Create an initial github workflow for InferenceSystem

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Fixes

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Fixes

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Install ffmpeg

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Fixes

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Fixes

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Add max iterations

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

* Remove docker test since won't be max iterations 1

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>

---------

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>