Improve precision of machine detection timestamp
Previously it was 1 minute granularity, which is still six
different .ts segments, some of which may have no predictions
(i.e., no calls).  With this fix, the best .ts segment
can be more easily identified since this fix uses the offset
in seconds of the best prediction within the minute to give
a more precise timestamp of a machine detection.

Fixes #292

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>